+++ logging/modulators/IterationModuloLoggingModulaor.go
@@ -7,11 +7,11 @@ import . "github.com/LindsayBradford/crm/annealing/shared"
 // IterationModuloLoggingModulator modulates FinishedIteration Annealing Event instances at a rate of 1 every modulo
 // events. StartedIteration events are completely filtered out. All other event types are allowed through to the LogHandler.
 type IterationModuloLoggingModulator struct {
-	iterationModulo uint
+	iterationModulo uint64
 }
 
 // WithModulo defines the modulo to apply against FinishedIteration Annealing Event instances.
-func (m *IterationModuloLoggingModulator) WithModulo(modulo uint) *IterationModuloLoggingModulator {
+func (m *IterationModuloLoggingModulator) WithModulo(modulo uint64) *IterationModuloLoggingModulator {
 	m.iterationModulo = modulo
 	return m
 }
